# public/swagger.json
{
	"components": {
...
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "rca-auth",
		"version": "1.0.0",
		"description": "RCA authentication system that helps user to authenticate base on RCA credentials",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Ntwari Clarance Liberiste"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/auth/current": {
			"get": {
				"operationId": "CurrentUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Retrieves the information of the current logged in user",
				"tags": [
					"Auth"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}